[Api::Filter] Refactor and/or expression arrays
Inverts the "if" statement to remove the use of `target`, which is
confusing to read.

Instead the "expr" is built first, and then based on the `:logical_or`
value, is put in the proper array.  This makes it just a bit easier to
logically parse as a human, but is functionally the same.