[Api::Filter] Allow filtering '=' via arrays
Adds functionality to the API to allow filtering a particular column by
one of many types of values.  This could be particularly useful when
filtering via a `type` column that is an STI column.

The SQL that would be generated from the filter would be the following:

    SELECT *
      FROM miq_requests
     WHERE (type = "MiqProvisionRequest" OR type = "VmMigrateRequest")

This avoids having to make changes to `MiqExpression` that would allow
for something similar to how `where(:type => [a,b,c])` works already,
and also ensures that we are using something that can be converted into
SQL on the backend.