[Performance] Add includes for direct relations
This change adds includes for direct relationships that are passed in as
attributes so they are loaded as part of the main query.

There is a quick RBac check just before allowing adding that attribute
to determine if it participates in rbac, but otherwise it works just
like an attribute.

The reason this didn't get included previously is relations like `tags`
and others are lumped into the `virtual_attributes_for` grouping for the
API, despite not being a virtual attribute (for "reasons"), and since it
is not a `virtual_includes` nor is it an attribute with a `dot`
notation, it doesn't get included with the rest.