introduce generic options parsing and error handling
the pattern for fetching update/create options for a resource are pretty consistent. pulled out into renderer

I didn't have a good way to call options' super so I left the if logic in the base controller. would have been nice to extract that.
will extract this into a number of other controllers once this has been merged