Allow for manual OPTIONS data validation
The 'data' key in the response body can be simple or complex.  You can now
opt-out of generic data validation in the complex case by not providing the
data.  This lets you manually validate the data in the specs that know what part
of the structure is important, while retaining the rest of the options
expectations.