add href back into multiple object errors
we are talking about results for multiple actions where there is an error
in one of the actions.

before before:
many of the responses had a href to the erroneous objects

before this PR:
the error result did not have a href to the object with an error

after this PR:
error results now have the href back in them.
Since we are consolidating on this block of code, this adds the urls back
to quite a few actions, and future wins, too.

ASIDE: Not thrilled with the verbosity of this fix.