pull out ems operations from enqueue_object_action to enqueue_ems_action
Some of our ems operations were passing a user and others were not.
This will now send user for all.

Since most of our calls are for ems operations, I was toying with the idea
of passing :role => nil for those that were general calls.
Decided a separate call was better.