Don't specify the resource when it is the controllers primary resource
There is no reason to override the type or specify the collection class that is
the controller's primary type.

I put a print statement next to ever one of these calls and verified that the
value coming in was the value I expected. So it verified that there were tests
for every method and the values were the expected controller model type.

The only issue found was 2 classes did not have tests. all others were successful
with the obvious values