consolidate api.yml put/delete/patch with post
Currently, we are not reading put or patch. They are part of post
Deleting the unused put and patch from api.yml makes it clearer

Currently we are reading both post/delete and delete/delete for collections
but only post/delete for resources.
Consolidating into post/delete makes it easier to have that more consistent