Extract Simplify api_action
- Move rescue blocks from api_resource to api_action
- api_action adds a href when the user passes in id=nil
- extracted logic to construct result hash if only the result_options was returned

note on full_action_results:

It would have made more logical sense to have the parameter be the action_result text message and the `&block` represent the program flow.

Instead, the parameter is the results of the original block yield, and the log message is passed in as the block.
This is similar to how logging works when passed a block.It lazy produces the string action_result text.