add queue_name for ems_operations
We queue ems operations to go to the ems operations worker.

Most of the places in our code we use `Model.method_queue` and the method handles the details.
A few places in the api use `enqueue_ems_action` to directly put the message on the queue.

enqueue_ems_action now properly sets the queue name

Some of our `ContainerImage` specs has a nil ems and were blowing up. So I added the `&.`