Reset session cache to fix sporadic ttl failures
We have been trying to track down this issue for a while.
This always works locally for me, and typically for api specs,
but this fails often for cross repo specs

Since the user token service is configured with the environment, and we
are altering the variables passed into the token service, I want to clear
out that service.

Changing the ttl values for non-auth tests, so I don't see the need to
clear across every test.