Fix cloud_volume clone supports spec
The spec was using stub_supports_not on the class, not the specific
cloud_volume which meant when autosde added clone support the spec began
to fail.

We should be using a generic factory not a provider specific factory,
and we should stub_supports_not on the object not just the class.