Fix specs requiring memcached
When memcached was not available, the specs would fail, but not in the
way that was expected. Before this commit, fetching a value would just
return nil, and while the spec would fail, it was misleading. Now, the
spec will fail early if memcached is not available and it is clear why
the spec fails.

Additionally, it was found that the closures in users_spec were invalid,
so while the spec was written to test sql, memory, and cache, due to the
closures all 3 sets of tests were actually running cache. This was more
evident when the memcached checks were added because both sql and memory
were both failing in the abscence of memcached. This commit fixes those
closures so they run accurately.