Filter#parse regexp to the rescue
Instead of going through each character to find the operators,
use a regular expression that is coded to handle all the operators

This assumes a filter is made up of a name, operator, value.
These assumptions seem to protect the code later on.

The original regular expression for the operator was: /[!<>=][=~]=?/
While easier to read, it had quite a few false positives.
The false positives were caught by the OPERATORS.key?, so not a big thing

I do wonder if we should be more greedy matching operators, and then
raising an error. The pedantic approach may make for a better developer
experience, but leaving it as it was for now.