Query metric rollups using start_date derived from the first rollup
This resolves a time based sporadic failure.  This means it works for many
hours of the day but fails with specific ones.

Previously, Time.zone is UTC, so Time.zone.today when run at
August 15th at 12:05 AM UTC would return Thu, 15 Aug 2024.

If we created the rollups using 1.hour.ago, the first one would be
August 14th 11:05 PM.

If we our rollups were on August 14th and we're looking for August 15th hourly
rollups, we find nothing.

This change ensures both the timestamp on the rollups and the start_date we
query the rollups with come from the same time.