Enforce valid offset and limit
Before
======

If the offset was invalid, it was just thrown away.
The Pen tests assumed they are able to hack the app.
They assumed we were vulnerable to a sql injection attack

After
=====

Throw an error if a bad value comes in for the limit or offset
Doesn't really make a difference, but makes the testers happy